src/main/java/dev/drf/leetcode/populating/next/right/pointers/in/each/node/PopulatingNextRightPointersInEachNodeProblem.java
package dev.drf.leetcode.populating.next.right.pointers.in.each.node;

import dev.drf.leetcode.Problem;

/**
 * https://leetcode.com/problems/populating-next-right-pointers-in-each-node/
 */
public class PopulatingNextRightPointersInEachNodeProblem implements Problem {
    public Node connect(Node root) {
        if (root == null) {
            return null;
        }
        if (root.left != null) {
            root.left.next = root.right;
        }
        if (root.right != null && root.next != null) {
            root.right.next = root.next.left;
        }

        connect(root.left);
        connect(root.right);

        return root;
    }
}
